chore: delete duplicate folders and files, move files in to kickflip-project folder (#8)
